+++ packages/at_client/lib/src/client/at_client_impl.dart
@@ -8,11 +8,15 @@ import 'package:at_client/src/client/local_secondary.dart';
 import 'package:at_client/src/client/remote_secondary.dart';
 import 'package:at_client/src/client/secondary.dart';
 import 'package:at_client/src/client/verb_builder_manager.dart';
+import 'package:at_client/src/compaction/at_commit_log_compaction.dart';
 import 'package:at_client/src/encryption_service/encryption_manager.dart';
+import 'package:at_client/src/listener/at_sign_change_listener.dart';
+import 'package:at_client/src/listener/switch_at_sign_event.dart';
 import 'package:at_client/src/manager/at_client_manager.dart';
 import 'package:at_client/src/manager/storage_manager.dart';
 import 'package:at_client/src/manager/sync_manager.dart';
 import 'package:at_client/src/manager/sync_manager_impl.dart';
+import 'package:at_client/src/preference/at_client_config.dart';
 import 'package:at_client/src/preference/at_client_preference.dart';
 import 'package:at_client/src/response/response.dart';
 import 'package:at_client/src/service/encryption_service.dart';
@@ -42,8 +46,11 @@ import 'package:meta/meta.dart';
 import 'package:path/path.dart';
 import 'package:uuid/uuid.dart';
 
-/// Implementation of [AtClient] interface
-class AtClientImpl implements AtClient {
+/// Implementation of [AtClient] interface and [AtSignChangeListener] interface
+///
+/// Implements to [AtSignChangeListener] to get notified on switch atSign event. On switch atSign event,
+/// pause's the compaction job on currentAtSign and start/resume the compaction job on the new atSign
+class AtClientImpl implements AtClient, AtSignChangeListener {
   AtClientPreference? _preference;
 
   AtClientPreference? get preference => _preference;
@@ -52,6 +59,11 @@ class AtClientImpl implements AtClient {
   SecondaryKeyStore? _localSecondaryKeyStore;
   LocalSecondary? _localSecondary;
   RemoteSecondary? _remoteSecondary;
+  AtClientCommitLogCompaction? _atClientCommitLogCompaction;
+  AtClientConfig? _atClientConfig;
+
+  AtClientCommitLogCompaction? get atClientCommitLogCompaction =>
+      _atClientCommitLogCompaction;
 
   @override
   // ignore: override_on_non_overriding_member
@@ -68,6 +80,7 @@ class AtClientImpl implements AtClient {
     _telemetry = telemetryService;
     _cascadeSetTelemetryService();
   }
+
   @override
   @experimental
   AtTelemetryService? get telemetry => _telemetry;
@@ -81,18 +94,22 @@ class AtClientImpl implements AtClient {
   AtChops? get atChops => _atChops;
 
   late SyncService _syncService;
+
   @override
-  set syncService (SyncService syncService) {
+  set syncService(SyncService syncService) {
     _syncService = syncService;
   }
+
   @override
   SyncService get syncService => _syncService;
 
   late NotificationService _notificationService;
+
   @override
-  set notificationService (NotificationService notificationService) {
+  set notificationService(NotificationService notificationService) {
     _notificationService = notificationService;
   }
+
   @override
   NotificationService get notificationService => _notificationService;
 
@@ -112,21 +129,31 @@ class AtClientImpl implements AtClient {
       RemoteSecondary? remoteSecondary,
       EncryptionService? encryptionService,
       SecondaryKeyStore? localSecondaryKeyStore,
-      AtChops? atChops}) async {
+      AtChops? atChops,
+      AtClientCommitLogCompaction? atClientCommitLogCompaction,
+      AtClientConfig? atClientConfig}) async {
+    atClientManager ??= AtClientManager.getInstance();
     currentAtSign = AtUtils.formatAtSign(currentAtSign)!;
+
+    // Fetch cached AtClientImpl for re-use, or create a new one and init it
+    AtClientImpl? atClientImpl;
     if (atClientInstanceMap.containsKey(currentAtSign)) {
-      return atClientInstanceMap[currentAtSign];
+      atClientImpl = atClientInstanceMap[currentAtSign];
+    } else {
+      atClientImpl = AtClientImpl._(
+          currentAtSign, namespace, preferences, atClientManager,
+          remoteSecondary: remoteSecondary,
+          encryptionService: encryptionService,
+          localSecondaryKeyStore: localSecondaryKeyStore,
+          atChops: atChops,
+          atClientCommitLogCompaction: atClientCommitLogCompaction,
+          atClientConfig: atClientConfig);
+
+      await atClientImpl._init();
     }
 
-    atClientManager ??= AtClientManager.getInstance();
-    var atClientImpl = AtClientImpl._(
-        currentAtSign, namespace, preferences, atClientManager,
-        remoteSecondary: remoteSecondary,
-        encryptionService: encryptionService,
-        localSecondaryKeyStore: localSecondaryKeyStore,
-        atChops: atChops);
-
-    await atClientImpl._init();
+    await atClientImpl!._startCompactionJob();
+    atClientManager.listenToAtSignChange(atClientImpl);
 
     atClientInstanceMap[currentAtSign] = atClientImpl;
     return atClientInstanceMap[currentAtSign];
@@ -137,7 +164,9 @@ class AtClientImpl implements AtClient {
       {RemoteSecondary? remoteSecondary,
       EncryptionService? encryptionService,
       SecondaryKeyStore? localSecondaryKeyStore,
-      AtChops? atChops}) {
+      AtChops? atChops,
+      AtClientCommitLogCompaction? atClientCommitLogCompaction,
+      AtClientConfig? atClientConfig}) {
     _atSign = AtUtils.formatAtSign(theAtSign)!;
     _logger = AtSignLogger('AtClientImpl ($_atSign)');
     _preference = preference;
@@ -152,6 +181,7 @@ class AtClientImpl implements AtClient {
     _remoteSecondary = remoteSecondary;
     _encryptionService = encryptionService;
     _atChops = atChops;
+    _atClientCommitLogCompaction = atClientCommitLogCompaction;
   }
 
   Future<void> _init() async {
@@ -166,7 +196,7 @@ class AtClientImpl implements AtClient {
 
     // Now using ??= because we may be injecting a RemoteSecondary
     _remoteSecondary ??= RemoteSecondary(_atSign, _preference!,
-        privateKey: _preference!.privateKey);
+        atChops: atChops, privateKey: _preference!.privateKey);
 
     // Now using ??= because we may be injecting an EncryptionService
     _encryptionService ??= EncryptionService(_atSign);
@@ -175,6 +205,23 @@ class AtClientImpl implements AtClient {
     _encryptionService!.localSecondary = _localSecondary;
 
     _cascadeSetTelemetryService();
+  }
+
+  Future<void> _startCompactionJob() async {
+    AtCompactionJob atCompactionJob = AtCompactionJob(
+        (await AtCommitLogManagerImpl.getInstance().getCommitLog(_atSign))!,
+        SecondaryPersistenceStoreFactory.getInstance()
+            .getSecondaryPersistenceStore(_atSign)!);
+
+    _atClientCommitLogCompaction ??=
+        AtClientCommitLogCompaction.create(_atSign, atCompactionJob);
+
+    _atClientConfig ??= AtClientConfig.getInstance();
+
+    if (!_atClientCommitLogCompaction!.isCompactionJobRunning()) {
+      _atClientCommitLogCompaction!.scheduleCompaction(
+          _atClientConfig!.commitLogCompactionTimeIntervalInMins);
+    }
   }
 
   /// Does nothing unless a telemetry service has been injected
@@ -212,12 +259,6 @@ class AtClientImpl implements AtClient {
   @override
   RemoteSecondary? getRemoteSecondary() {
     return _remoteSecondary;
-  }
-
-  @override
-  @Deprecated("Use SyncManager.sync")
-  SyncManager? getSyncManager() {
-    return SyncManagerImpl.getInstance().getSyncManager(_atSign);
   }
 
   @override
@@ -616,7 +657,8 @@ class AtClientImpl implements AtClient {
     var command =
         'stream:init$sharedWith namespace:$namespace $streamId $fileName ${encryptedData.length}\n';
     _logger.finer('sending stream init:$command');
-    var remoteSecondary = RemoteSecondary(_atSign, _preference!);
+    var remoteSecondary =
+        RemoteSecondary(_atSign, _preference!, atChops: atChops);
     var result = await remoteSecondary.executeCommand(command, auth: true);
     _logger.finer('ack message:$result');
     if (result != null && result.startsWith('stream:ack')) {
@@ -701,8 +743,7 @@ class AtClientImpl implements AtClient {
           ..metadata!.ttl = 2592000000
           ..sharedBy = _atSign;
 
-        var notificationResult =
-            await notificationService.notify(
+        var notificationResult = await notificationService.notify(
           NotificationParams.forUpdate(
             atKey,
             value: jsonEncode(fileTransferObject.toJson()),
@@ -829,11 +870,6 @@ class AtClientImpl implements AtClient {
     }
   }
 
-  @Deprecated("Use EncryptionService")
-  Future<void> encryptUnEncryptedData() async {
-    await _encryptionService!.encryptUnencryptedData();
-  }
-
   @override
   String? getCurrentAtSign() => _atSign;
 
@@ -943,47 +979,21 @@ class AtClientImpl implements AtClient {
     return await getRemoteSecondary()?.executeVerb(builder);
   }
 
-  //
-  // Everything after this point has been deprecated
-  //
-
-  /// Returns a new instance of [AtClient]. App has to pass the current user atSign
-  /// and the client preference.
-  @Deprecated("Use AtClientManger to get instance of atClient")
-  static Future<AtClient?> getClient(String? currentAtSign) async {
-    if (atClientInstanceMap.containsKey(currentAtSign)) {
-      return atClientInstanceMap[currentAtSign];
+  @override
+  void listenToAtSignChange(SwitchAtSignEvent switchAtSignEvent) {
+    // Checks if the instance of AtClientImpl belongs to previous atSign. If Yes,
+    // the compaction job is stopped and removed from changeListener list.
+    if (switchAtSignEvent.previousAtClient?.getCurrentAtSign() ==
+        getCurrentAtSign()) {
+      _atClientCommitLogCompaction!.stopCompactionJob();
+      _atClientManager.removeChangeListeners(this);
     }
-    AtSignLogger('AtClientImpl').severe(
-        'Instance of AtClientImpl for $currentAtSign has not been created');
-    return null;
   }
 
-  @Deprecated("Use [create]")
-
-  /// use [create]
-  static Future<void> createClient(String currentAtSign, String? namespace,
-      AtClientPreference preferences) async {
-    currentAtSign = AtUtils.formatAtSign(currentAtSign)!;
-    if (atClientInstanceMap.containsKey(currentAtSign)) {
-      return;
-    }
-    if (preferences.isLocalStoreRequired) {
-      var storageManager = StorageManager(preferences);
-      await storageManager.init(currentAtSign, preferences.keyStoreSecret);
-    }
-    var atClientImpl = AtClientImpl(currentAtSign, namespace, preferences);
-    await atClientImpl._init();
-    atClientInstanceMap[currentAtSign] = atClientImpl;
-  }
-
-  @Deprecated("Use [create]")
-  AtClientImpl(
-      String atSign,
-      String? namespace,
-      AtClientPreference preference) {
-    _atSign = AtUtils.formatAtSign(atSign)!;
-    _preference = preference;
-    _namespace = namespace;
+  // TODO v4 - remove this method in version 4 of at_client package
+  @override
+  @Deprecated("Use AtClient.syncService")
+  SyncManager? getSyncManager() {
+    return SyncManagerImpl.getInstance().getSyncManager(_atSign);
   }
 }