+++ packages/at_client/lib/src/client/at_client_spec.dart
@@ -20,7 +20,7 @@ abstract class AtClient {
   /// Returns a singleton instance of [SyncManager] that is responsible for syncing data between
   /// local secondary server and remote secondary server.
   /// [Deprecated] Use [AtClient.syncService]
-  @Deprecated("Use SyncManager.sync")
+  @Deprecated("Use AtClient.syncService")
   SyncManager? getSyncManager();
 
   @experimental
@@ -38,10 +38,10 @@ abstract class AtClient {
 
   AtChops? get atChops;
 
-  set syncService (SyncService syncService);
+  set syncService(SyncService syncService);
   SyncService get syncService;
 
-  set notificationService (NotificationService notificationService);
+  set notificationService(NotificationService notificationService);
   NotificationService get notificationService;
 
   /// Sets the preferences such as sync strategy, storage path etc., for the client.