+++ packages/at_client/lib/src/manager/at_client_manager.dart
@@ -1,3 +1,4 @@
+import 'package:at_chops/at_chops.dart';
 import 'package:at_client/at_client.dart';
 import 'package:at_client/src/listener/at_sign_change_listener.dart';
 import 'package:at_client/src/listener/switch_at_sign_event.dart';
@@ -26,7 +27,10 @@ class AtClientManager {
 
   AtClient get atClient => _currentAtClient!;
 
+  @Deprecated('Use atClientManager.atClient.syncService')
   SyncService get syncService => atClient.syncService;
+
+  @Deprecated('Use atClientManager.atClient.notificationService')
   NotificationService get notificationService => atClient.notificationService;
 
   SecondaryAddressFinder? secondaryAddressFinder;
@@ -50,36 +54,43 @@ class AtClientManager {
     }
   }
 
+  /// * Can provide an [AtServiceFactory] to control what types of AtClients, NotificationServices and SyncServices get created
+  /// by this method.
+  /// * Can provide an [AtChops] instance if available; setCurrentAtSign will ensure that it is injected into objects that
+  /// can use it
   Future<AtClientManager> setCurrentAtSign(
-      String atSign, String? namespace, AtClientPreference preference) async {
+      String atSign, String? namespace, AtClientPreference preference,
+      {AtServiceFactory? serviceFactory, AtChops? atChops}) async {
+    serviceFactory ??= DefaultAtServiceFactory();
+
     _logger.info("setCurrentAtSign called with atSign $atSign");
     AtUtils.fixAtSign(atSign);
     secondaryAddressFinder ??= CacheableSecondaryAddressFinder(
         preference.rootDomain, preference.rootPort);
     if (_currentAtClient != null &&
         _currentAtClient?.getCurrentAtSign() == atSign) {
-      _logger.info('previous currentAtSign ${_currentAtClient?.getCurrentAtSign()} is same as new atSign $atSign - doing nothing, returning');
+      _logger.info(
+          'previous currentAtSign ${_currentAtClient?.getCurrentAtSign()} is same as new atSign $atSign - doing nothing, returning');
       return this;
     }
 
-    _logger.info('Switching atSigns from ${_currentAtClient?.getCurrentAtSign()} to $atSign');
+    _logger.info(
+        'Switching atSigns from ${_currentAtClient?.getCurrentAtSign()} to $atSign');
     _atSign = atSign;
     var previousAtClient = _currentAtClient;
-    _currentAtClient = await AtClientImpl.create(_atSign, namespace, preference,
-        atClientManager: this);
+    _currentAtClient = await serviceFactory
+        .atClient(_atSign, namespace, preference, this, atChops: atChops);
+
     final switchAtSignEvent =
         SwitchAtSignEvent(previousAtClient, _currentAtClient!);
     _notifyListeners(switchAtSignEvent);
 
-    var notificationService = await NotificationServiceImpl.create(
-        _currentAtClient!,
-        atClientManager: this);
+    var notificationService =
+        await serviceFactory.notificationService(_currentAtClient!, this);
     _currentAtClient!.notificationService = notificationService;
 
-    var syncService = await SyncServiceImpl.create(
-        _currentAtClient!,
-        atClientManager: this,
-        notificationService: notificationService);
+    var syncService = await serviceFactory.syncService(
+        _currentAtClient!, this, notificationService);
     _currentAtClient!.syncService = syncService;
 
     _logger.info("setCurrentAtSign complete");
@@ -88,7 +99,7 @@ class AtClientManager {
   }
 
   void listenToAtSignChange(AtSignChangeListener listener) {
-    if (! _changeListeners.contains(listener)) {
+    if (!_changeListeners.contains(listener)) {
       _changeListeners.add(listener);
     }
   }
@@ -97,7 +108,8 @@ class AtClientManager {
     // Copying the items in _changeListener to a new list to avoid
     // concurrent modification exception when removing the previous
     // atSign listeners
-    List<AtSignChangeListener> copyOfChangeListeners = List.from(_changeListeners);
+    List<AtSignChangeListener> copyOfChangeListeners =
+        List.from(_changeListeners);
     for (var listener in copyOfChangeListeners) {
       listener.listenToAtSignChange(switchAtSignEvent);
     }
@@ -119,5 +131,50 @@ class AtClientManager {
   @visibleForTesting
   Iterator<dynamic> getItemsInChangeListeners() {
     return _changeListeners.iterator;
+  }
+
+  /// NOT A PART of API. Added for unit tests
+  @visibleForTesting
+  removeAllChangeListeners() {
+    _changeListeners.clear();
+  }
+}
+
+abstract class AtServiceFactory {
+  Future<AtClient> atClient(String atSign, String? namespace,
+      AtClientPreference preference, AtClientManager atClientManager,
+      {AtChops? atChops});
+
+  Future<NotificationService> notificationService(
+      AtClient atClient, AtClientManager atClientManager);
+
+  Future<SyncService> syncService(AtClient atClient,
+      AtClientManager atClientManager, NotificationService notificationService);
+}
+
+class DefaultAtServiceFactory implements AtServiceFactory {
+  @override
+  Future<AtClient> atClient(String atSign, String? namespace,
+      AtClientPreference preference, AtClientManager atClientManager,
+      {AtChops? atChops}) async {
+    return await AtClientImpl.create(atSign, namespace, preference,
+        atClientManager: atClientManager, atChops: atChops);
+  }
+
+  @override
+  Future<NotificationService> notificationService(
+      AtClient atClient, AtClientManager atClientManager) async {
+    return await NotificationServiceImpl.create(atClient,
+        atClientManager: atClientManager);
+  }
+
+  @override
+  Future<SyncService> syncService(
+      AtClient atClient,
+      AtClientManager atClientManager,
+      NotificationService notificationService) async {
+    return await SyncServiceImpl.create(atClient,
+        atClientManager: atClientManager,
+        notificationService: notificationService);
   }
 }