+++ packages/at_client/test/samples/monitor/monitor_retry_force_close_socket.dart
@@ -7,7 +7,9 @@ void main() async {
     final atClientManager = await AtClientManager.getInstance()
         .setCurrentAtSign(aliceAtSign, 'wavi', TestUtil.getAlicePreference());
 
-    atClientManager.atClient.notificationService.subscribe().listen((notification) {
+    atClientManager.atClient.notificationService
+        .subscribe()
+        .listen((notification) {
       _notificationCallback(notification);
     });
     atClientManager.atClient.notificationService
@@ -16,8 +18,10 @@ void main() async {
       _notificationCallback(notification);
     });
     print('stopping monitor');
-    Future.delayed(Duration(seconds: 5),
-        () => atClientManager.atClient.notificationService.stopAllSubscriptions());
+    Future.delayed(
+        Duration(seconds: 5),
+        () => atClientManager.atClient.notificationService
+            .stopAllSubscriptions());
   } on Exception catch (e, trace) {
     print(e.toString());
     print(trace);